Connecting MCTS with Key Events and Improving CapturePoint

+++ Source/GameAI_Project/Public/Actors/AmmoCrate.h
@@ -50,7 +50,7 @@ public:
 
 	/** Only allow collection if ammo is below this percentage */
 	UPROPERTY(EditAnywhere, Category = "AmmoCrate")
-	int32 MaxAmmoPercentForCollection = 0.99f;
+	float MaxAmmoPercentForCollection = 0.99f;
 
 	/** Play sound on collection? */
 	UPROPERTY(EditAnywhere, Category = "AmmoCrate|Audio")
